+++ src/machinable/engine/dry_engine.py
@@ -12,9 +12,8 @@ class DryEngine(Engine):
     def supports_resources():
         return False
 
-    def storage_middleware(self, storage):
-        storage["url"] = "mem://"
-        return storage
+    def on_before_storage_creation(self, execution):
+        execution.storage.config["url"] = "mem://"
 
     def _submit(self, execution):
         return execution